warning 제거

+++ Unreal10th_HealloWorld/Orc.cpp
@@ -3,6 +3,6 @@
 void Orc::Skill(Actor* InTarget) const
 {
 	printf("[%s] : 도끼 던지기 스킬을 사용합니다.\n", Name.c_str());
-	int Damage = CalcDamage() * 3.0f;
+	int Damage = static_cast<int>(CalcDamage() * 3.0f);
 	ApplyDamage(InTarget, Damage);
 }
